Made a note on non overlapping x
As it is currently written, non overlapping x will confuse on what return values it expects from students, project tests will check they return a Dataframe and a Series, the instructions ask to return a 2d ndarray and a 1d ndarray. Not sure which one is right and wrong though so just commented here.